requirements.txt
python-dotenv
discord.py
pydantic